refactor(relay): C1 identity 타입을 stream_relay/identity.rs로 추출 — giant-file 임계 회피
stream_relay.rs 1014→981 prod LoC (<1000). re-export로 기존 import 경로 유지.
인벤토리 문서 재생성 포함. 동작 불변.

Implemented-by: Codex (session 019f2547-98ea-7fd0-98f0-e9fd509df50f)
