+++ src/services/cluster/stream_relay.rs
@@ -54,6 +54,9 @@ use tokio::task::JoinHandle;
 
 use super::session_matcher::MatchedChannel;
 
+mod identity;
+pub use identity::{RelayDroppedFrame, RelayTurnIdentity};
+
 /// Default size of the producer → relay queue. Generous enough to absorb a
 /// burst of provider output (e.g. a long planning block dumping thousands of
 /// lines at once) without losing data, bounded so a stuck consumer cannot
@@ -96,42 +99,6 @@ pub enum DeliveryOutcome {
     Delivered,
     NotDelivered,
     Unknown,
-}
-
-/// The turn identity stamped on a relayed frame. Terminal frames use this as the
-/// commit-fence identity gate; non-terminal frames may also carry it so producer
-/// backpressure can attribute an evicted frame to the affected turn.
-#[derive(Clone, Debug, Default, PartialEq, Eq)]
-pub struct RelayTurnIdentity {
-    pub turn_user_msg_id: u64,
-    pub turn_started_at: String,
-    pub turn_start_offset: Option<u64>,
-}
-
-impl RelayTurnIdentity {
-    pub fn has_strict_turn_start_offset(&self) -> bool {
-        self.turn_start_offset.is_some()
-    }
-}
-
-/// A frame evicted from the producer queue before the relay task consumed it.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct RelayDroppedFrame {
-    pub sequence: u64,
-    pub turn_identity: RelayTurnIdentity,
-}
-
-impl RelayDroppedFrame {
-    fn from_frame(frame: StreamFrame) -> Self {
-        Self {
-            sequence: frame.sequence,
-            turn_identity: RelayTurnIdentity {
-                turn_user_msg_id: frame.turn_user_msg_id,
-                turn_started_at: frame.turn_started_at,
-                turn_start_offset: frame.turn_start_offset,
-            },
-        }
-    }
 }
 
 /// An opaque stream frame emitted by a provider. Carries enough metadata for
